split up revision.tsx
